rgw: asio: don't parse the rest of request if invalid
Currently for requests that contain data like PUT, even though we send a HTTP
status to the client we still read the data from the socket, which is set to
keep-alive, so we're effectively reading the next request if the client chooses
to send this and hence reading invalid data. This should fix a few s3-tests
where requests after a expected failure continue to fail as beast would throw an
invalid method.

Fixes: https://tracker.ceph.com/issues/42208
Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>