rgw: close connections for PUT/POST requests when aborting early
This avoids the frontends to read the invalid data from the socket to determine
whether more data is sent and determining whether content-length amount of
data was already sent by a client before reading the next connection. Forcing
the client to close will allow us not to discard the remaining data as we're not
sure whether the client will send the rest of the data or not. Most clients will
not send anymore data if they send the expect header, which means the data we
are discarding could be the next request, making further requests parse
incorrectly.

Fixes: https://tracker.ceph.com/issues/42208
Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>