crush: ftotal needn't to be set to 0 again in every replicas loop
If we set ftotal to 0 in every loop begins, r may has the same value
as previous iteration, which is meaningless. The previous iteration
may has already reject this r due to collision or item out, so in
this iteration the same r will be reject again.
We should keep r' always increasing to avoid redudant calculation.

In Sage Weil's paper "CRUSH: Controlled, Scalable, Decentralized
Placement of Replicated Data", ftotal is set to 0 only once before
loop begins. (Algorithm 1 CRUSH placement for object x, line 7).
So I think this should be a mistake.

Signed-off-by: Gao Mingfei <mingfei.gao@ucloud.cn>