ceph-volume: allow non-physical devices for "lvm prepare"
Currently only physical devices and partitions can be passed to
"lvm prepare --data </path/to/device>".
This relaxes the limitation by allowing any devices except for LVs.
LVs shoud be passed as `--data <vg_name>/<lv_name>`.

Multipath devices has not been supported on purpose.
https://docs.ceph.com/docs/nautilus/ceph-volume/lvm/prepare/#multipath-support

The concern was that it's hard to work `ceph-volume` correctly with many
multipath scenarios.  However, it's up to the admin to make sure
that multipath is set up as expected.

As of other kind of devices, there have not been prohibited explicitly.
In addition, admin is in charge of configuraing these devices too.

Fixes: https://tracker.ceph.com/issues/42502
Signed-off-by: Satoru Takeuchi <sat@cybozu.co.jp>