mgr/telemetry: send device telemetry via per-host POST to device endpoint
We do not want to associate devices with clusters because that may
communicate unnecessary information about the association between vendors
and clusters (which, when large, are potentially identifying).

Instead, do a POST per host with all of the devices on that host only.

The devices endpoint does not log the POST time, so these per-host
records won't be associated with each other.

Signed-off-by: Sage Weil <sage@redhat.com>