log: just return if t is empty
When the t is empty, there is no need to do the followed t.clear
or the m_log_buf flush. Because the m_log_buf will be fed by _flush()
only when t is not empty and then immediately it will be flushed to
log files and emptied just before returned from _flush().

In other word, m_log_buf just a buffer for the log messages in _flush()
before flushing them to log files.

This will also remove the 'requeue' pararmeter since in all the cases
we should requeue the new entry.

Signed-off-by: Xiubo Li <xiubli@redhat.com>