mgr/diskprediction_local: Updated dependencies in ceph.spec.in, debian/control to match requirements.txt
Added pandas dependency to ceph.spec.in and debian/control.

In the spirit of "if it aint broke, dont fix it", I did NOT add
scikit-learn as a dependency in spec or control, because scikit-learn
was already a dependency in diskprediction_local, and so it should have
already have been taken care of.

Also in the same spirit, removed joblib dependency from requirements.txt
because scikit-learn depends on it and therefore joblib will get
installed when scikit-learn gets installed.

Signed-off-by: Karanraj Chauhan <kachauha@redhat.com>