mon/MonMap: encode (more) valid compat monmap when we have v2-only addrs
If we have 1 or more mons with v2-only addrs, pre-nautilus clients can't
talk to them.  If there are more than 1 such mons in the map, they also
fail when loading the map because they expect the addrs to be unique. In
such situations, lie by giving them v1 addrs that are actually v2 ip:port
(so not actually valid).  Hopefully there are enough other mons that do
have v1 addrs that the clients can still connect.

Fixes: https://tracker.ceph.com/issues/42600
Signed-off-by: Sage Weil <sage@redhat.com>