Revert "mgr/balancer: add per pool concurrent optimization limit"
The commit message for the original commit didn't seem to match the
code.

It isn't clear how the count of non-active+clean PGs relates
to how many optimizations should be done.

In addition this change wasn't reducing the max value to
calc_pg_upmaps() as changes are made by each call.  Thus
invalidating the previous commit message that included
"...upmap_max_optimizations will remain a hard limit for the
total optimizations that can be returned."

This reverts commit 86444dbfe5d8478530182116507d034d9e180f5e.

Signed-off-by: David Zafman <dzafman@redhat.com>