RGW: url-encode the kms keyid in reqeusts to barbican
Barbican operates via a REST api and we pass it the keyid as a part of
the URL. As such we need to make sure the keyid is URL encoded in case
the keyid ever uses a character not supported HTTP protocol.

This patch encodes the keyid sent to barbican by using the url_encode
method from rgw_common. A new `url_encode` signiture with a sting_view
source has been added to help keep the code simple.

Fixes: https://tracker.ceph.com/issues/39488
Signed-off-by: Matthew Oliver <moliver@suse.com>