mgr/ssh: implement mon add, mgr add
If you currently want to increase the number of instances of a
mgr or mon you have to use:

`ceph orchestartor {mon, mgr} update $current_num+1 $host:$network_conf`

This is fine for rook, but not so userfriendly for the SSH-Orchestrator.

Adding an more imperative way of adding and removing mons/mgrs.

`ceph orchestrator {mon, mgr} {add, rm} $host:$optional_network_conf`

Signed-off-by: Joshua Schmid <jschmid@suse.de>