test/msgr: fix ComplexTest fail when using DPDK protocal stack
The DPDK protocol stack failed to run ComplexTest. When testing the DPDK
protocol stack, the same session will be received by the same queue by the
RSS. The test case ComplexTest uses two CPUs to run the server and client
respectively. The message sent by the client cannot be forwarded to the
server, so the test fails. When support_local_listen_table() is true,
let the client connect to the server on the same CPU and the test pass.

Signed-off-by: Chunsong Feng <fengchunsong@huawei.com>