mgr/dashboard: Add binary min and max validators
Adds simple validators that resemble the min and max validators for
numbers.

As a binary size (the string) has first be converted into bytes
(the number) to validate against some other number and in case of an
error the other number want's to be converted into a size to display,
the whole process needs a bunch of conversion, but these validators take
care of all that.

Fixes: https://tracker.ceph.com/issues/38287
Signed-off-by: Stephan Müller <smueller@suse.com>