cmake: Set missing dependencies
When building Windows DLLs, all symbols must be resolved. For this
reason, we'll have to make sure that we don't have any unresolved
symbols.

We ended up having some issues with multiple definitions when
doing this for Linux targets, so we'll do this only for Windows
buids for now: http://paste.openstack.org/raw/787049/

We'll avoid having the librbd plugins as a librbd dependency since
this introduces a circular dependency.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>
Signed-off-by: Alin Gabriel Serdean <aserdean@cloudbasesolutions.com>