ceph_json: json encoding filter
Can create and register a filter (that will be pointed at by the formatter
as an external feature), that could be then used to replace default
encoders for specific data types. This can be useful in cases where
we want to mutate the output (for example: translate zone ids to zone
names on the fly).

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>