rgw: Replace boost::string_view with std::string_view
std::string_view interacts better with the standard library overall
and we can expect new code to use it.

Keeping both would be nice, unfortunately if we have an overload for
boost::string_view and one for std::string_view any call with
std::string will fail as ambiguous, so they don't go together well.

In theory it would be nice to define ceph::string_view as an alias,
but because we have all those awful 'using namespace std' and 'using
namespace ceph' all over, it'll clash.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>