mon: annotate more tell commands
currently, the only consumer of the annotated tell commands is
`Monitor::handle_tell_command()`, which check the cap of the client
against the required permission of the command to be executed.

if the command is not annotated, it's assumed that the command requires
"mon rwx".

because the asock commands are registered at runtime using
`AdminSocket::register_command()`, and there are some asock commands are
shared by all ceph applications using asock facility, without a proper
machinary, we cannot annotate all of them and populate this annotated
commands to all these ceph applications. so, at this moment, only
a subset of asock commands are annotated.

Signed-off-by: Kefu Chai <kchai@redhat.com>