mgr/volumes: add support for ssh-orchestrator
added parameter for host (optional)

old:
ceph fs volume create <fs_name>

new:
ceph fs volume create <fs_name> (hostname)

This will allow the ssh-orchestrator to deploy
a MDS daemon on the provided host.
If this is not specified and the ssh-backend is used
it will result in an RuntimeError. The volumes will
still be created though.

Signed-off-by: Joshua Schmid <jschmid@suse.de>