mgr/dashboard: Avoid escaped quotes using TSLint
The problem was that the TSLint linting hasn't matched the
linting of prettier regarding the quotes of a string.

Explanation:
'\'' will be converted to "'" with prettier, but TSLint thrown an
error that it wanted the string to look like '\''.

Now TSLint will not error if it sees an single quote inside a double
quote to match the behavior of prettier.

Fixes: https://tracker.ceph.com/issues/38287
Signed-off-by: Stephan Müller <smueller@suse.com>