os/bluestore/KernelDevice: get rid of 'nvme' type
We are either 'hdd' or 'ssd' based on the rotational flag.  Previously,
we would try to distinguish between an nvme vs SATA/SAS ssd and set the
class to 'nvme'.  This was misguided--the interface is not important and
has no direct bearing on the device performance.  Moreover, the HDD
manufacturers are planning to produce rotation HDDs that use the nvme
interface instead of SATA/SAS.

So, drop this.

This may be somewhat disruptive to clusters where we were detecting
nvme but now are not.  However, the good news is that this doesn't seem
to trigger for real deployments because LVM breaks the is_nvme()
method.

Signed-off-by: Sage Weil <sage@redhat.com>