mon: elector: spread current election scores in election messages
This has two benefits:
1) when we defer to a peer X, we can't defer to a different peer Y
unless X will also defer to Y. So we keep those scores around and
check when we get proposals. (Note that we might defer to XYZ in order,
and X may not defer to Z -- but that should be okay as X will defer to
Y, and as Y will defer to Z, nobody wins and the election cycles!)
2) We spread our current score view on every message so that if the
election cycles, everybody should have the newest (and SAME) scores
to run from, so it converges on the next go-round.

Add a memo-ized ConnectionTracker::get_encoded_bl() so we don't
need to repeatedly encode when  sendng out proposals, and use
it throughout Elector.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>
(cherry picked from commit 334066f01eae360cc35d46beb952adb8dfad3482)

Conflicts:
	src/messages/MMonElection.h

Resolves: rhbz#1800382