suite: add --filter-all parameter
Add --filter-all parameter to 'teuthology-suite' to allow
precise test case selection. The --filter-in makes use of filter list
with logical 'or' operation to include test cases.
The --filter-out takes filter list and excludes any test case which
match any from the filter list.

This patch provides with 'and' logical operation for a given filter list.

Signed-off-by: Kyr Shatskyy <kyrylo.shatskyy@suse.com>