osd/SnapMirror: no need to record purged_snaps every epoch
These keys were a mirror of the mon's purged_epoch keys.  We were writing
them, but never looking at them.  We only need the per-pool keys that
are indexed by snapid for lookups, so we should stop storing the per-epoch
records.

Note that the OSDSuperblock has the purged_snaps_last that serves as our
position/cursor for ingesting the mon keys.

Fixes: https://tracker.ceph.com/issues/42012
Signed-off-by: Sage Weil <sage@redhat.com>