rpm: explicitly declare python-tox build dependency
"make check" needs python-tox to run, but until 71ac2163831ffd764bf3da6a3efa76ef02e5e884
("ceph.spec: added dashboard_v2 development and runtime dependencies")
this was not explicitly declared in the spec file. It does not make sense to
backport that commit to luminous, though, because the dashboard itself was never
backported to luminous.

Fixes: https://tracker.ceph.com/issues/43082
Signed-off-by: Nathan Cutler <ncutler@suse.com>