doc: add Python 2 to Ubuntu 18.04 installations
This commit updates the "Create a Cluster" section of the
"Storage Cluster Quick Start" document. Ubuntu 18.04 does
not have Python 2 installed by default. This commit instructs
users of Python 18.04 to install the package "python-minimal"
to provide Python 2 prior to running "ceph-deploy".

This commit fixes Bug number 9 in the list here:
https://pad.ceph.com/p/Report_Documentation_Bugs

Signed-off-by: Zac Dover <zac.dover@gmail.com>