ceph-detect-init: run tox tests on Python 2 only
Luminous is EOL upstream and 12.2.13 will be the last point release, so it's
pretty much a given that luminous will not be migrated to py3.

Luminous runs on OS versions where py3 is not a "first-class citizen": compiling
the mgr with a py3 subinterpreter on such systems would be "interesting"...

Python 3 testing was dropped for ceph-disk by
28c545eff5f73e392b70ef4c0cf74f8a896b693b (over 3 years ago) and
ceph-detect-init is a component of ceph-disk.

This commit cannot be cherry-picked from master because ceph-detect-init was
removed in nautilus.

Fixes: https://tracker.ceph.com/issues/43100
Signed-off-by: Nathan Cutler <ncutler@suse.com>