common/config: update values when they are removed via mon
If a value is set via the mon (ceph config set ...), and then later
removed, the removal doesn't propagate to the legacy values (or debug
options) because update_legacy_values() only iterates over legacy value
and not debug options.

Similarly, we want to trigger a config notification to any observers,
so we need to add the option to values.changed.

Fixes: https://tracker.ceph.com/issues/42964
Signed-off-by: Sage Weil <sage@redhat.com>