doc: s/#/$ sudo/ -- working around Sphinx's policy on hashes (#)
This commit changes "#" to "$ sudo", because when hash (#) appears
near the beginning of a line in an .rst file and then that file is
parsed by Sphinx, the material after the hash appears blue and in
italics. This blue, italicized text is undesriable and it is my
opinion that it might lead a reader to think that the authors of
the document are trying to impart some kind of information by
means of the color and italics. So: goobye hashes, hello sudo
commands.

Signed-off-by: Zac Dover <zac.dover@gmail.com>