rgw: create_bucket(): reset read_version
We might get a non empty objv_tracker.read_version when we try to create
a bucket. This could happen if bucket was created on different zone and
we got its meta passed in. Need to clear version, otherwise it will fail
when we try to write the bucket info.

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>