cephadm: do not force container removal on unit stop
The -f arg for 'podman rm' will remove the container without actually
stopping it, leaving the running containers and processes (conmon and
daemon) running and generally making a mess of things.

Doing 'podman rm' without the -f on a running container will complain
that the container is still running.

So, first stop, then remove.

Fixes: 48ab80cdaa1fe959be729e671233c11e4df1a5d1
Signed-off-by: Sage Weil <sage@redhat.com>