log: print pthread ID / name mapping in recent events dump.
For investigation of a multithreading issue it might be
beneficial to obtain the name of a thread responsible
for generating a given log entry. This can be easily
accomplished with e.g. `info threads` of GDB when core
dump is available. However, that's not always the case.

This change adds a new section to the Ceph's "recent
events" log dump with the mapping between pthread's ID
and its name for all threads referenced in the recent
entries buffer.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>